Extract section "Stripe Connect Usage"
The part about running requests on behalf of a connected account came
too early in the README, before the regular "Usage" section.

I preferred to introduce it at a later point.

Also, I prefixed the calls to the scripts in this section with './'
to clarify that scripts are called through their relative location
starting at the root of the repository.